+++ src/game/WorldHandlers/SpellAuras.cpp
@@ -2172,1293 +2172,13 @@ void Aura::TriggerSpellWithValue()
 /***                  AURA EFFECTS                     ***/
 /*********************************************************/
 
-/**
- * @brief Applies or removes a mounted display from the target.
- *
- * @param apply True to mount; false to unmount.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraMounted(bool apply, bool Real)
-{
-    // only at real add/remove aura
-    if (!Real)
-    {
-        return;
-    }
-
-    Unit* target = GetTarget();
-
-    if (apply)
-    {
-        CreatureInfo const* ci = ObjectMgr::GetCreatureTemplate(m_modifier.m_miscvalue);
-        if (!ci)
-        {
-            sLog.outErrorDb("AuraMounted: `creature_template`='%u' not found in database (only need it modelid)", m_modifier.m_miscvalue);
-            return;
-        }
-
-        uint32 display_id = Creature::ChooseDisplayId(ci);
-        CreatureModelInfo const* minfo = sObjectMgr.GetCreatureModelRandomGender(display_id);
-        if (minfo)
-        {
-            display_id = minfo->modelid;
-        }
-
-        target->Mount(display_id, GetId());
-    }
-    else
-    {
-        target->Unmount(true);
-    }
-}
-
-/**
- * @brief Applies or removes water walking on the target.
- *
- * @param apply True to enable water walking; false to disable it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraWaterWalk(bool apply, bool Real)
-{
-    // only at real add/remove aura
-    if (!Real)
-    {
-        return;
-    }
-
-    GetTarget()->SetWaterWalk(apply);
-}
-
-/**
- * @brief Applies or removes feather fall on the target.
- *
- * @param apply True to enable feather fall; false to disable it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraFeatherFall(bool apply, bool Real)
-{
-    // only at real add/remove aura
-    if (!Real)
-    {
-        return;
-    }
-
-    GetTarget()->SetFeatherFall(apply);
-}
-
-/**
- * @brief Applies or removes hovering movement state.
- *
- * @param apply True to enable hover; false to disable it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraHover(bool apply, bool Real)
-{
-    // only at real add/remove aura
-    if (!Real)
-    {
-        return;
-    }
-
-    GetTarget()->SetHover(apply);
-}
-
-/**
- * @brief Refreshes client breathing timers for the target.
- */
-void Aura::HandleWaterBreathing(bool /*apply*/, bool /*Real*/)
-{
-    // update timers in client
-    if (GetTarget()->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)GetTarget())->UpdateMirrorTimers();
-    }
-}
-
-/**
- * @brief Applies or removes a shapeshift form and its related state changes.
- *
- * @param apply True to enter the form; false to leave it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraModShapeshift(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    ShapeshiftForm form = ShapeshiftForm(m_modifier.m_miscvalue);
-
-    SpellShapeshiftFormEntry const* ssEntry = sSpellShapeshiftFormStore.LookupEntry(form);
-    if (!ssEntry)
-    {
-        sLog.outError("Unknown shapeshift form %u in spell %u", form, GetId());
-        return;
-    }
-
-    uint32 modelid = 0;
-    Powers PowerType = POWER_MANA;
-    Unit* target = GetTarget();
-
-    // remove SPELL_AURA_EMPATHY
-    target->RemoveSpellsCausingAura(SPELL_AURA_EMPATHY);
-
-    if (ssEntry->modelID_A)
-    {
-        // i will asume that creatures will always take the defined model from the dbc
-        // since no field in creature_templates describes wether an alliance or
-        // horde modelid should be used at shapeshifting
-        if (target->GetTypeId() != TYPEID_PLAYER)
-        {
-            modelid = ssEntry->modelID_A;
-        }
-        else
-        {
-            // players are a bit different since the dbc has seldomly an horde modelid
-            if (Player::TeamForRace(target->getRace()) == HORDE)
-            {
-                // get model for race ( in 2.2.4 no horde models in dbc field, only 0 in it
-                modelid = sObjectMgr.GetModelForRace(ssEntry->modelID_A, target->getRaceMask());
-            }
-
-            // nothing found in above, so use default
-            if (!modelid)
-            {
-                modelid = ssEntry->modelID_A;
-            }
-        }
-    }
-
-    // remove polymorph before changing display id to keep new display id
-    switch (form)
-    {
-        case FORM_CAT:
-        case FORM_TREE:
-        case FORM_TRAVEL:
-        case FORM_AQUA:
-        case FORM_BEAR:
-        case FORM_DIREBEAR:
-        case FORM_FLIGHT_EPIC:
-        case FORM_FLIGHT:
-        case FORM_MOONKIN:
-        {
-            // remove movement affects
-            target->RemoveSpellsCausingAura(SPELL_AURA_MOD_ROOT, GetHolder());
-            Unit::AuraList const& slowingAuras = target->GetAurasByType(SPELL_AURA_MOD_DECREASE_SPEED);
-            for (Unit::AuraList::const_iterator iter = slowingAuras.begin(); iter != slowingAuras.end();)
-            {
-                SpellEntry const* aurSpellInfo = (*iter)->GetSpellProto();
-
-                uint32 aurMechMask = GetAllSpellMechanicMask(aurSpellInfo);
-
-                // If spell that caused this aura has Croud Control or Daze effect
-                if ((aurMechMask & MECHANIC_NOT_REMOVED_BY_SHAPESHIFT) ||
-                    // some Daze spells have these parameters instead of MECHANIC_DAZE (skip snare spells)
-                    (aurSpellInfo->SpellIconID == 15 && aurSpellInfo->Dispel == 0 &&
-                     (aurMechMask & (1 << (MECHANIC_SNARE - 1))) == 0))
-                {
-                    ++iter;
-                    continue;
-                }
-
-                // All OK, remove aura now
-                target->RemoveAurasDueToSpellByCancel(aurSpellInfo->Id);
-                iter = slowingAuras.begin();
-            }
-
-            // and polymorphic affects
-            if (target->IsPolymorphed())
-            {
-                target->RemoveAurasDueToSpell(target->GetTransform());
-            }
-
-            //no break here
-        }
-        case FORM_GHOSTWOLF:
-        {
-            // remove water walk aura. TODO:: there is probably better way to do this
-            target->RemoveSpellsCausingAura(SPELL_AURA_WATER_WALK);
-
-            break;
-        }
-        default:
-            break;
-    }
-
-    if (apply)
-    {
-        // remove other shapeshift before applying a new one
-        target->RemoveSpellsCausingAura(SPELL_AURA_MOD_SHAPESHIFT, GetHolder());
-
-        if (modelid > 0)
-        {
-            target->SetObjectScale(DEFAULT_OBJECT_SCALE);
-            target->SetDisplayId(modelid);
-        }
-
-        // now only powertype must be set
-        switch (form)
-        {
-            case FORM_CAT:
-                PowerType = POWER_ENERGY;
-                break;
-            case FORM_BEAR:
-            case FORM_DIREBEAR:
-            case FORM_BATTLESTANCE:
-            case FORM_BERSERKERSTANCE:
-            case FORM_DEFENSIVESTANCE:
-                PowerType = POWER_RAGE;
-                break;
-            default:
-                break;
-        }
-
-        if (PowerType != POWER_MANA)
-        {
-            // reset power to default values only at power change
-            if (target->GetPowerType() != PowerType)
-            {
-                target->SetPowerType(PowerType);
-            }
-
-            switch (form)
-            {
-                case FORM_CAT:
-                case FORM_BEAR:
-                case FORM_DIREBEAR:
-                {
-                    // get furor proc chance
-                    int32 furorChance = 0;
-                    Unit::AuraList const& mDummy = target->GetAurasByType(SPELL_AURA_DUMMY);
-                    for (Unit::AuraList::const_iterator i = mDummy.begin(); i != mDummy.end(); ++i)
-                    {
-                        if ((*i)->GetSpellProto()->SpellIconID == 238)
-                        {
-                            furorChance = (*i)->GetModifier()->m_amount;
-                            break;
-                        }
-                    }
-
-                    if (m_modifier.m_miscvalue == FORM_CAT)
-                    {
-                        target->SetPower(POWER_ENERGY, 0);
-                        if (irand(1, 100) <= furorChance)
-                        {
-                            target->CastSpell(target, 17099, true, NULL, this);
-                        }
-                    }
-                    else
-                    {
-                        target->SetPower(POWER_RAGE, 0);
-                        if (irand(1, 100) <= furorChance)
-                        {
-                            target->CastSpell(target, 17057, true, NULL, this);
-                        }
-                    }
-                    break;
-                }
-                case FORM_BATTLESTANCE:
-                case FORM_DEFENSIVESTANCE:
-                case FORM_BERSERKERSTANCE:
-                {
-                    uint32 Rage_val = 0;
-                    // Stance mastery + Tactical mastery (both passive, and last have aura only in defense stance, but need apply at any stance switch)
-                    if (target->GetTypeId() == TYPEID_PLAYER)
-                    {
-                        PlayerSpellMap const& sp_list = ((Player*)target)->GetSpellMap();
-                        for (PlayerSpellMap::const_iterator itr = sp_list.begin(); itr != sp_list.end(); ++itr)
-                        {
-                            if (itr->second.state == PLAYERSPELL_REMOVED)
-                            {
-                                continue;
-                            }
-
-                            SpellEntry const* spellInfo = sSpellStore.LookupEntry(itr->first);
-                            if (spellInfo && spellInfo->SpellFamilyName == SPELLFAMILY_WARRIOR && spellInfo->SpellIconID == 139)
-                            {
-                                Rage_val += target->CalculateSpellDamage(target, spellInfo, EFFECT_INDEX_0) * 10;
-                            }
-                        }
-                    }
-
-                    if (target->GetPower(POWER_RAGE) > Rage_val)
-                    {
-                        target->SetPower(POWER_RAGE, Rage_val);
-                    }
-                    break;
-                }
-                default:
-                    break;
-            }
-        }
-
-        target->SetShapeshiftForm(form);
-
-        // a form can give the player a new castbar with some spells.. this is a clientside process..
-        // serverside just needs to register the new spells so that player isn't kicked as cheater
-        if (target->GetTypeId() == TYPEID_PLAYER)
-            for (uint32 i = 0; i < 8; ++i)
-            {
-                if (ssEntry->spellId[i])
-                {
-                    ((Player*)target)->addSpell(ssEntry->spellId[i], true, false, false, false);
-                }
-            }
-    }
-    else
-    {
-        if (modelid > 0)
-        {
-            target->SetDisplayId(target->GetNativeDisplayId());
-        }
-
-        if (target->getClass() == CLASS_DRUID)
-        {
-            target->SetPowerType(POWER_MANA);
-        }
-
-        target->SetShapeshiftForm(FORM_NONE);
-
-        switch (form)
-        {
-                // Nordrassil Harness - bonus
-            case FORM_BEAR:
-            case FORM_DIREBEAR:
-            case FORM_CAT:
-                if (Aura* dummy = target->GetDummyAura(37315))
-                {
-                    target->CastSpell(target, 37316, true, NULL, dummy);
-                }
-                break;
-                // Nordrassil Regalia - bonus
-            case FORM_MOONKIN:
-                if (Aura* dummy = target->GetDummyAura(37324))
-                {
-                    target->CastSpell(target, 37325, true, NULL, dummy);
-                }
-                break;
-            default:
-                break;
-        }
-
-        // look at the comment in apply-part
-        if (target->GetTypeId() == TYPEID_PLAYER)
-            for (uint32 i = 0; i < 8; ++i)
-            {
-                if (ssEntry->spellId[i])
-                {
-                    ((Player*)target)->removeSpell(ssEntry->spellId[i], false, false, false);
-                }
-            }
-    }
-
-    // adding/removing linked auras
-    // add/remove the shapeshift aura's boosts
-    HandleShapeshiftBoosts(apply);
-
-    if (target->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)target)->InitDataForForm();
-    }
-}
-
-/**
- * @brief Applies or removes a transform model effect.
- *
- * @param apply True to apply the transform; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraTransform(bool apply, bool Real)
-{
-    Unit* target = GetTarget();
-    if (apply)
-    {
-        // special case (spell specific functionality)
-        if (m_modifier.m_miscvalue == 0)
-        {
-            switch (GetId())
-            {
-                case 16739:                                 // Orb of Deception
-                {
-                    uint32 orb_model = target->GetNativeDisplayId();
-                    switch (orb_model)
-                    {
-                            // Troll Female
-                        case 1479: target->SetDisplayId(10134); break;
-                            // Troll Male
-                        case 1478: target->SetDisplayId(10135); break;
-                            // Tauren Male
-                        case 59:   target->SetDisplayId(10136); break;
-                            // Human Male
-                        case 49:   target->SetDisplayId(10137); break;
-                            // Human Female
-                        case 50:   target->SetDisplayId(10138); break;
-                            // Orc Male
-                        case 51:   target->SetDisplayId(10139); break;
-                            // Orc Female
-                        case 52:   target->SetDisplayId(10140); break;
-                            // Dwarf Male
-                        case 53:   target->SetDisplayId(10141); break;
-                            // Dwarf Female
-                        case 54:   target->SetDisplayId(10142); break;
-                            // NightElf Male
-                        case 55:   target->SetDisplayId(10143); break;
-                            // NightElf Female
-                        case 56:   target->SetDisplayId(10144); break;
-                            // Undead Female
-                        case 58:   target->SetDisplayId(10145); break;
-                            // Undead Male
-                        case 57:   target->SetDisplayId(10146); break;
-                            // Tauren Female
-                        case 60:   target->SetDisplayId(10147); break;
-                            // Gnome Male
-                        case 1563: target->SetDisplayId(10148); break;
-                            // Gnome Female
-                        case 1564: target->SetDisplayId(10149); break;
-                            // BloodElf Female
-                        case 15475: target->SetDisplayId(17830); break;
-                            // BloodElf Male
-                        case 15476: target->SetDisplayId(17829); break;
-                            // Dranei Female
-                        case 16126: target->SetDisplayId(17828); break;
-                            // Dranei Male
-                        case 16125: target->SetDisplayId(17827); break;
-                        default: break;
-                    }
-                    break;
-                }
-                case 42365:                                 // Murloc costume
-                    target->SetDisplayId(21723);
-                    break;
-                    // case 44186:                          // Gossip NPC Appearance - All, Brewfest
-                    // break;
-                    // case 48305:                          // Gossip NPC Appearance - All, Spirit of Competition
-                    // break;
-                case 50517:                                 // Dread Corsair
-                case 51926:                                 // Corsair Costume
-                {
-                    // expected for players
-                    uint32 race = target->getRace();
-
-                    switch (race)
-                    {
-                        case RACE_HUMAN:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25037 : 25048);
-                            break;
-                        case RACE_ORC:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25039 : 25050);
-                            break;
-                        case RACE_DWARF:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25034 : 25045);
-                            break;
-                        case RACE_NIGHTELF:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25038 : 25049);
-                            break;
-                        case RACE_UNDEAD:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25042 : 25053);
-                            break;
-                        case RACE_TAUREN:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25040 : 25051);
-                            break;
-                        case RACE_GNOME:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25035 : 25046);
-                            break;
-                        case RACE_TROLL:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25041 : 25052);
-                            break;
-                        case RACE_GOBLIN:                   // not really player race (3.x), but model exist
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25036 : 25047);
-                            break;
-                        case RACE_BLOODELF:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25032 : 25043);
-                            break;
-                        case RACE_DRAENEI:
-                            target->SetDisplayId(target->getGender() == GENDER_MALE ? 25033 : 25044);
-                            break;
-                    }
-
-                    break;
-                }
-                // case 50531:                              // Gossip NPC Appearance - All, Pirate Day
-                // break;
-                // case 51010:                              // Dire Brew
-                // break;
-                default:
-                    sLog.outError("Aura::HandleAuraTransform, spell %u does not have creature entry defined, need custom defined model.", GetId());
-                    break;
-            }
-        }
-        else                                                // m_modifier.m_miscvalue != 0
-        {
-            uint32 model_id;
-
-            CreatureInfo const* ci = ObjectMgr::GetCreatureTemplate(m_modifier.m_miscvalue);
-            if (!ci)
-            {
-                model_id = 16358;                           // pig pink ^_^
-                sLog.outError("Auras: unknown creature id = %d (only need its modelid) Form Spell Aura Transform in Spell ID = %d", m_modifier.m_miscvalue, GetId());
-            }
-            else
-            {
-                model_id = Creature::ChooseDisplayId(ci);    // Will use the default model here
-            }
-
-            target->SetDisplayId(model_id);
-
-            // creature case, need to update equipment if additional provided
-            if (ci && target->GetTypeId() == TYPEID_UNIT)
-            {
-                ((Creature*)target)->LoadEquipment(ci->EquipmentTemplateId, false);
-            }
-
-            // Dragonmaw Illusion (set mount model also)
-            if (GetId() == 42016 && target->GetMountID() && !target->GetAurasByType(SPELL_AURA_MOD_FLIGHT_SPEED_MOUNTED).empty())
-            {
-                target->SetUInt32Value(UNIT_FIELD_MOUNTDISPLAYID, 16314);
-            }
-        }
-
-        // update active transform spell only not set or not overwriting negative by positive case
-        if (!target->GetTransform() || !IsPositiveSpell(GetId()) || IsPositiveSpell(target->GetTransform()))
-        {
-            target->SetTransform(GetId());
-        }
-
-        // polymorph case
-        if (Real && target->GetTypeId() == TYPEID_PLAYER && target->IsPolymorphed())
-        {
-            // for players, start regeneration after 1s (in polymorph fast regeneration case)
-            // only if caster is Player (after patch 2.4.2)
-            if (GetCasterGuid().IsPlayer())
-            {
-                ((Player*)target)->setRegenTimer(1 * IN_MILLISECONDS);
-            }
-
-            // dismount polymorphed target (after patch 2.4.2)
-            if (target->IsMounted())
-            {
-                target->RemoveSpellsCausingAura(SPELL_AURA_MOUNTED, GetHolder());
-            }
-        }
-    }
-    else                                                    // !apply
-    {
-        // ApplyModifier(true) will reapply it if need
-        target->SetTransform(0);
-        target->SetDisplayId(target->GetNativeDisplayId());
-
-        // apply default equipment for creature case
-        if (target->GetTypeId() == TYPEID_UNIT)
-        {
-            ((Creature*)target)->LoadEquipment(((Creature*)target)->GetCreatureInfo()->EquipmentTemplateId, true);
-        }
-
-        // re-apply some from still active with preference negative cases
-        Unit::AuraList const& otherTransforms = target->GetAurasByType(SPELL_AURA_TRANSFORM);
-        if (!otherTransforms.empty())
-        {
-            // look for other transform auras
-            Aura* handledAura = *otherTransforms.begin();
-            for (Unit::AuraList::const_iterator i = otherTransforms.begin(); i != otherTransforms.end(); ++i)
-            {
-                // negative auras are preferred
-                if (!IsPositiveSpell((*i)->GetSpellProto()->Id))
-                {
-                    handledAura = *i;
-                    break;
-                }
-            }
-            handledAura->ApplyModifier(true);
-        }
-
-        // Dragonmaw Illusion (restore mount model)
-        if (GetId() == 42016 && target->GetMountID() == 16314)
-        {
-            if (!target->GetAurasByType(SPELL_AURA_MOUNTED).empty())
-            {
-                uint32 cr_id = target->GetAurasByType(SPELL_AURA_MOUNTED).front()->GetModifier()->m_miscvalue;
-                if (CreatureInfo const* ci = ObjectMgr::GetCreatureTemplate(cr_id))
-                {
-                    uint32 display_id = Creature::ChooseDisplayId(ci);
-                    CreatureModelInfo const* minfo = sObjectMgr.GetCreatureModelRandomGender(display_id);
-                    if (minfo)
-                    {
-                        display_id = minfo->modelid;
-                    }
-
-                    target->SetUInt32Value(UNIT_FIELD_MOUNTDISPLAYID, display_id);
-                }
-            }
-        }
-    }
-}
-
-/**
- * @brief Applies or removes a forced reputation reaction for a player.
- *
- * @param apply True to apply the forced reaction; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleForceReaction(bool apply, bool Real)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (!Real)
-    {
-        return;
-    }
-
-    Player* player = (Player*)GetTarget();
-
-    uint32 faction_id = m_modifier.m_miscvalue;
-    ReputationRank faction_rank = ReputationRank(m_modifier.m_amount);
-
-    player->GetReputationMgr().ApplyForceReaction(faction_id, faction_rank, apply);
-    player->GetReputationMgr().SendForceReactions();
-
-    // stop fighting if at apply forced rank friendly or at remove real rank friendly
-    if ((apply && faction_rank >= REP_FRIENDLY) || (!apply && player->GetReputationRank(faction_id) >= REP_FRIENDLY))
-    {
-        player->StopAttackFaction(faction_id);
-    }
-}
-
-/**
- * @brief Applies or removes a player skill bonus from the aura.
- *
- * @param apply True to apply the bonus; false to remove it.
- */
-void Aura::HandleAuraModSkill(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    uint32 prot = GetSpellProto()->EffectMiscValue[m_effIndex];
-    int32 points = GetModifier()->m_amount;
-
-    ((Player*)GetTarget())->ModifySkillBonus(prot, (apply ? points : -points), m_modifier.m_auraname == SPELL_AURA_MOD_SKILL_TALENT);
-    if (prot == SKILL_DEFENSE)
-    {
-        ((Player*)GetTarget())->UpdateDefenseBonusesMod();
-    }
-}
-
-/**
- * @brief Awards the configured item when a channel-death aura ends by death.
- *
- * @param apply True on application; false on removal.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleChannelDeathItem(bool apply, bool Real)
-{
-    if (Real && !apply)
-    {
-        if (m_removeMode != AURA_REMOVE_BY_DEATH)
-        {
-            return;
-        }
-        // Item amount
-        if (m_modifier.m_amount <= 0)
-        {
-            return;
-        }
-
-        SpellEntry const* spellInfo = GetSpellProto();
-        if (spellInfo->EffectItemType[m_effIndex] == 0)
-        {
-            return;
-        }
-
-        Unit* victim = GetTarget();
-        Unit* caster = GetCaster();
-        if (!caster || caster->GetTypeId() != TYPEID_PLAYER)
-        {
-            return;
-        }
-
-        // Soul Shard (target req.)
-        if (spellInfo->EffectItemType[m_effIndex] == 6265)
-        {
-            // Only from non-grey units
-            if (!((Player*)caster)->isHonorOrXPTarget(victim) ||
-                (victim->GetTypeId() == TYPEID_UNIT && !((Creature*)victim)->IsTappedBy((Player*)caster)))
-            {
-                return;
-            }
-        }
-
-        // Adding items
-        uint32 noSpaceForCount = 0;
-        uint32 count = m_modifier.m_amount;
-
-        ItemPosCountVec dest;
-        InventoryResult msg = ((Player*)caster)->CanStoreNewItem(NULL_BAG, NULL_SLOT, dest, spellInfo->EffectItemType[m_effIndex], count, &noSpaceForCount);
-        if (msg != EQUIP_ERR_OK)
-        {
-            count -= noSpaceForCount;
-            ((Player*)caster)->SendEquipError(msg, NULL, NULL, spellInfo->EffectItemType[m_effIndex]);
-            if (count == 0)
-            {
-                return;
-            }
-        }
-
-        Item* newitem = ((Player*)caster)->StoreNewItem(dest, spellInfo->EffectItemType[m_effIndex], true);
-        ((Player*)caster)->SendNewItem(newitem, count, true, true);
-    }
-}
-
-/**
- * @brief Redirects the caster camera to the target while the aura is active.
- *
- * @param apply True to bind sight; false to restore normal view.
- */
-void Aura::HandleBindSight(bool apply, bool /*Real*/)
-{
-    Unit* caster = GetCaster();
-    if (!caster || caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Camera& camera = ((Player*)caster)->GetCamera();
-    if (apply)
-    {
-        camera.SetView(GetTarget());
-    }
-    else
-    {
-        camera.ResetView();
-    }
-}
-
-/**
- * @brief Redirects the caster camera for farsight while the aura is active.
- *
- * @param apply True to enable farsight; false to restore normal view.
- */
-void Aura::HandleFarSight(bool apply, bool /*Real*/)
-{
-    Unit* caster = GetCaster();
-    if (!caster || caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Camera& camera = ((Player*)caster)->GetCamera();
-    if (apply)
-    {
-        camera.SetView(GetTarget());
-    }
-    else
-    {
-        camera.ResetView();
-    }
-}
-
-/**
- * @brief Applies or removes creature tracking flags on a player.
- *
- * @param apply True to enable tracking; false to disable it.
- */
-void Aura::HandleAuraTrackCreatures(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        GetTarget()->RemoveNoStackAurasDueToAuraHolder(GetHolder());
-    }
-
-    if (apply)
-    {
-        GetTarget()->SetFlag(PLAYER_TRACK_CREATURES, uint32(1) << (m_modifier.m_miscvalue - 1));
-    }
-    else
-    {
-        GetTarget()->RemoveFlag(PLAYER_TRACK_CREATURES, uint32(1) << (m_modifier.m_miscvalue - 1));
-    }
-}
-
-/**
- * @brief Applies or removes resource tracking flags on a player.
- *
- * @param apply True to enable tracking; false to disable it.
- */
-void Aura::HandleAuraTrackResources(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        GetTarget()->RemoveNoStackAurasDueToAuraHolder(GetHolder());
-    }
-
-    if (apply)
-    {
-        GetTarget()->SetFlag(PLAYER_TRACK_RESOURCES, uint32(1) << (m_modifier.m_miscvalue - 1));
-    }
-    else
-    {
-        GetTarget()->RemoveFlag(PLAYER_TRACK_RESOURCES, uint32(1) << (m_modifier.m_miscvalue - 1));
-    }
-}
-
-/**
- * @brief Applies or removes stealthed-unit tracking on a player.
- *
- * @param apply True to enable tracking; false to disable it.
- */
-void Aura::HandleAuraTrackStealthed(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        GetTarget()->RemoveNoStackAurasDueToAuraHolder(GetHolder());
-    }
-
-    GetTarget()->ApplyModByteFlag(PLAYER_FIELD_BYTES, 0, PLAYER_FIELD_BYTE_TRACK_STEALTHED, apply);
-}
-
-/**
- * @brief Applies or removes a scale modifier and refreshes model data.
- *
- * @param apply True to apply the scale change; false to remove it.
- */
-void Aura::HandleAuraModScale(bool apply, bool /*Real*/)
-{
-    GetTarget()->ApplyPercentModFloatValue(OBJECT_FIELD_SCALE_X, float(m_modifier.m_amount), apply);
-    GetTarget()->UpdateModelData();
-}
-
 /*********************************************************/
 /***                     IMMUNITY                      ***/
 /*********************************************************/
 
-/**
- * @brief Applies special proc-trigger spell setup for specific aura spells.
- *
- * @param apply True to apply the proc aura; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraProcTriggerSpell(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    Unit* target = GetTarget();
-
-    switch (GetId())
-    {
-            // some spell have charges by functionality not have its in spell data
-        case 28200:                                         // Ascendance (Talisman of Ascendance trinket)
-            if (apply)
-            {
-                GetHolder()->SetAuraCharges(6);
-            }
-            break;
-        default:
-            break;
-    }
-}
-
-/**
- * @brief Applies or removes the tracked-unit dynamic flag.
- *
- * @param apply True to mark the unit as tracked; false to clear it.
- */
-void Aura::HandleAuraModStalked(bool apply, bool /*Real*/)
-{
-    // used by spells: Hunter's Mark, Mind Vision, Syndicate Tracker (MURP) DND
-    if (apply)
-    {
-        GetTarget()->SetFlag(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_TRACK_UNIT);
-    }
-    else
-    {
-        GetTarget()->RemoveFlag(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_TRACK_UNIT);
-    }
-}
-
 /*********************************************************/
 /***                   PERIODIC                        ***/
 /*********************************************************/
-
-void Aura::HandlePeriodicTriggerSpell(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-
-    Unit* target = GetTarget();
-
-    if (!apply)
-    {
-        switch (GetId())
-        {
-            case 66:                                        // Invisibility
-                if (m_removeMode == AURA_REMOVE_BY_EXPIRE)
-                {
-                    target->CastSpell(target, 32612, true, NULL, this);
-                }
-
-                return;
-            case 29213:                                     // Curse of the Plaguebringer
-                if (m_removeMode != AURA_REMOVE_BY_DISPEL)
-                    // Cast Wrath of the Plaguebringer if not dispelled
-                {
-                    target->CastSpell(target, 29214, true, 0, this);
-                }
-                return;
-            case 42783:                                     // Wrath of the Astrom...
-                if (m_removeMode == AURA_REMOVE_BY_EXPIRE && GetEffIndex() + 1 < MAX_EFFECT_INDEX)
-                {
-                    target->CastSpell(target, GetSpellProto()->CalculateSimpleValue(SpellEffectIndex(GetEffIndex() + 1)), true);
-                }
-
-                return;
-            default:
-                break;
-        }
-    }
-}
-
-/**
- * @brief Enables or disables periodic trigger handling with an explicit value.
- *
- * @param apply True to enable periodic processing; false to disable it.
- */
-void Aura::HandlePeriodicTriggerSpellWithValue(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-}
-
-/**
- * @brief Enables or disables periodic energize processing.
- *
- * @param apply True to enable periodic processing; false to disable it.
- */
-void Aura::HandlePeriodicEnergize(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-}
-
-/**
- * @brief Enables or disables periodic power burn processing.
- *
- * @param apply True to enable periodic processing; false to disable it.
- */
-void Aura::HandleAuraPowerBurn(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-}
-
-void Aura::HandleAuraPeriodicDummy(bool apply, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    Unit* target = GetTarget();
-
-    // For prevent double apply bonuses
-    bool loading = (target->GetTypeId() == TYPEID_PLAYER && ((Player*)target)->GetSession()->PlayerLoading());
-
-    SpellEntry const* spell = GetSpellProto();
-    switch (spell->SpellFamilyName)
-    {
-        case SPELLFAMILY_ROGUE:
-        {
-            // Master of Subtlety
-            if (spell->Id == 31666 && !apply)
-            {
-                target->RemoveAurasDueToSpell(31665);
-                break;
-            }
-            break;
-        }
-        case SPELLFAMILY_HUNTER:
-        {
-            // Aspect of the Viper
-            if (spell->SpellFamilyFlags & UI64LIT(0x0004000000000000))
-            {
-                // Update regen on remove
-                if (!apply && target->GetTypeId() == TYPEID_PLAYER)
-                {
-                    ((Player*)target)->UpdateManaRegen();
-                }
-                break;
-            }
-            break;
-        }
-    }
-
-    m_isPeriodic = apply;
-}
-
-/**
- * @brief Enables periodic healing and precalculates healing bonuses when applied.
- *
- * @param apply True to enable periodic healing; false to disable it.
- */
-void Aura::HandlePeriodicHeal(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-
-    Unit* target = GetTarget();
-
-    // For prevent double apply bonuses
-    bool loading = (target->GetTypeId() == TYPEID_PLAYER && ((Player*)target)->GetSession()->PlayerLoading());
-
-    // Custom damage calculation after
-    if (apply)
-    {
-        if (loading)
-        {
-            return;
-        }
-
-        Unit* caster = GetCaster();
-        if (!caster)
-        {
-            return;
-        }
-
-        m_modifier.m_amount = caster->SpellHealingBonusDone(target, GetSpellProto(), m_modifier.m_amount, DOT, GetStackAmount());
-    }
-}
-
-/**
- * @brief Enables periodic damage and precalculates damage bonuses when applied.
- *
- * @param apply True to enable periodic damage; false to disable it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandlePeriodicDamage(bool apply, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    m_isPeriodic = apply;
-
-    Unit* target = GetTarget();
-    SpellEntry const* spellProto = GetSpellProto();
-
-    // For prevent double apply bonuses
-    bool loading = (target->GetTypeId() == TYPEID_PLAYER && ((Player*)target)->GetSession()->PlayerLoading());
-
-    // Custom damage calculation after
-    if (apply)
-    {
-        if (loading)
-        {
-            return;
-        }
-
-        Unit* caster = GetCaster();
-        if (!caster)
-        {
-            return;
-        }
-
-        switch (spellProto->SpellFamilyName)
-        {
-            case SPELLFAMILY_WARRIOR:
-            {
-                // Rend
-                if (spellProto->SpellFamilyFlags & UI64LIT(0x0000000000000020))
-                {
-                    // 0.00743*(($MWB+$mwb)/2+$AP/14*$MWS) bonus per tick
-                    float ap = caster->GetTotalAttackPowerValue(BASE_ATTACK);
-                    int32 mws = caster->GetAttackTime(BASE_ATTACK);
-                    float mwb_min = caster->GetWeaponDamageRange(BASE_ATTACK, MINDAMAGE);
-                    float mwb_max = caster->GetWeaponDamageRange(BASE_ATTACK, MAXDAMAGE);
-                    m_modifier.m_amount += int32(((mwb_min + mwb_max) / 2 + ap * mws / 14000) * 0.00743f);
-                }
-                break;
-            }
-            case SPELLFAMILY_DRUID:
-            {
-                // Rip
-                if (spellProto->SpellFamilyFlags & UI64LIT(0x000000000000800000))
-                {
-                    if (caster->GetTypeId() != TYPEID_PLAYER)
-                    {
-                        break;
-                    }
-
-                    // $AP * min(0.06*$cp, 0.24)/6 [Yes, there is no difference, whether 4 or 5 CPs are being used]
-
-                    uint8 cp = ((Player*)caster)->GetComboPoints();
-
-                    // Idol of Feral Shadows. Cant be handled as SpellMod in SpellAura:Dummy due its dependency from CPs
-                    Unit::AuraList const& dummyAuras = caster->GetAurasByType(SPELL_AURA_DUMMY);
-                    for (Unit::AuraList::const_iterator itr = dummyAuras.begin(); itr != dummyAuras.end(); ++itr)
-                    {
-                        if ((*itr)->GetId() == 34241)
-                        {
-                            m_modifier.m_amount += cp * (*itr)->GetModifier()->m_amount;
-                            break;
-                        }
-                    }
-
-                        if (cp > 4)
-                        {
-                            cp = 4;
-                        }
-                        m_modifier.m_amount += int32(caster->GetTotalAttackPowerValue(BASE_ATTACK) * cp / 100);
-                }
-                break;
-            }
-            case SPELLFAMILY_ROGUE:
-            {
-                // Rupture
-                if (spellProto->SpellFamilyFlags & UI64LIT(0x000000000000100000))
-                {
-                    if (caster->GetTypeId() != TYPEID_PLAYER)
-                    {
-                        break;
-                    }
-                    // Dmg/tick = $AP*min(0.01*$cp, 0.03) [Like Rip: only the first three CP increase the contribution from AP]
-                    uint8 cp = ((Player*)caster)->GetComboPoints();
-                    if (cp > 3)
-                    {
-                        cp = 3;
-                    }
-                    m_modifier.m_amount += int32(caster->GetTotalAttackPowerValue(BASE_ATTACK) * cp / 100);
-                }
-                break;
-            }
-            default:
-                break;
-        }
-
-        if (m_modifier.m_auraname == SPELL_AURA_PERIODIC_DAMAGE)
-        {
-            // SpellDamageBonusDone for magic spells
-            if (spellProto->DmgClass == SPELL_DAMAGE_CLASS_NONE || spellProto->DmgClass == SPELL_DAMAGE_CLASS_MAGIC)
-            {
-                m_modifier.m_amount = caster->SpellDamageBonusDone(target, GetSpellProto(), m_modifier.m_amount, DOT, GetStackAmount());
-            }
-            // MeleeDamagebonusDone for weapon based spells
-            else
-            {
-                WeaponAttackType attackType = GetWeaponAttackType(GetSpellProto());
-                m_modifier.m_amount = caster->MeleeDamageBonusDone(target, m_modifier.m_amount, attackType, GetSpellProto(), DOT, GetStackAmount());
-            }
-        }
-    }
-    // remove time effects
-    else
-    {
-        // Parasitic Shadowfiend - handle summoning of two Shadowfiends on DoT expire
-        if (spellProto->Id == 41917)
-        {
-            target->CastSpell(target, 41915, true);
-        }
-    }
-}
-
-/**
- * @brief Enables or disables periodic percentage-based damage processing.
- *
- * @param apply True to enable periodic processing; false to disable it.
- */
-void Aura::HandlePeriodicDamagePCT(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-}
-
-/**
- * @brief Enables periodic health leech and precalculates spell bonuses when applied.
- *
- * @param apply True to enable periodic leeching; false to disable it.
- */
-void Aura::HandlePeriodicLeech(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-
-    // For prevent double apply bonuses
-    bool loading = (GetTarget()->GetTypeId() == TYPEID_PLAYER && ((Player*)GetTarget())->GetSession()->PlayerLoading());
-
-    // Custom damage calculation after
-    if (apply)
-    {
-        if (loading)
-        {
-            return;
-        }
-
-        Unit* caster = GetCaster();
-        if (!caster)
-        {
-            return;
-        }
-
-        m_modifier.m_amount = caster->SpellDamageBonusDone(GetTarget(), GetSpellProto(), m_modifier.m_amount, DOT, GetStackAmount());
-    }
-}
-
-/**
- * @brief Enables or disables periodic mana leech processing.
- *
- * @param apply True to enable periodic processing; false to disable it.
- */
-void Aura::HandlePeriodicManaLeech(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-}
-
-/**
- * @brief Enables periodic health funnel processing and precalculates bonuses when applied.
- *
- * @param apply True to enable periodic processing; false to disable it.
- */
-void Aura::HandlePeriodicHealthFunnel(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-
-    // For prevent double apply bonuses
-    bool loading = (GetTarget()->GetTypeId() == TYPEID_PLAYER && ((Player*)GetTarget())->GetSession()->PlayerLoading());
-
-    // Custom damage calculation after
-    if (apply)
-    {
-        if (loading)
-        {
-            return;
-        }
-
-        Unit* caster = GetCaster();
-        if (!caster)
-        {
-            return;
-        }
-
-        m_modifier.m_amount = caster->SpellDamageBonusDone(GetTarget(), GetSpellProto(), m_modifier.m_amount, DOT, GetStackAmount());
-    }
-}
 
 /*********************************************************/
 /***                  MODIFY STATS                     ***/
@@ -3468,586 +2188,13 @@ void Aura::HandlePeriodicHealthFunnel(bool apply, bool /*Real*/)
 /***        RESISTANCE        ***/
 /********************************/
 
-void Aura::HandleAuraModResistanceExclusive(bool apply, bool /*Real*/)
-{
-    for (int8 x = SPELL_SCHOOL_NORMAL; x < MAX_SPELL_SCHOOL; ++x)
-    {
-        if (m_modifier.m_miscvalue & int32(1 << x))
-        {
-            GetTarget()->HandleStatModifier(UnitMods(UNIT_MOD_RESISTANCE_START + x), BASE_VALUE, float(m_modifier.m_amount), apply);
-            if (GetTarget()->GetTypeId() == TYPEID_PLAYER)
-            {
-                ((Player*)GetTarget())->ApplyResistanceBuffModsMod(SpellSchools(x), m_positive, float(m_modifier.m_amount), apply);
-            }
-        }
-    }
-}
-
-/**
- * @brief Applies or removes flat resistance modifiers.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleAuraModResistance(bool apply, bool /*Real*/)
-{
-    Unit* target = GetTarget();
-
-    for (int8 x = SPELL_SCHOOL_NORMAL; x < MAX_SPELL_SCHOOL; ++x)
-    {
-        if (m_modifier.m_miscvalue & int32(1 << x))
-        {
-            target->HandleStatModifier(UnitMods(UNIT_MOD_RESISTANCE_START + x), TOTAL_VALUE, float(m_modifier.m_amount), apply);
-            if (target->GetTypeId() == TYPEID_PLAYER || ((Creature*)target)->IsPet())
-            {
-                ((Player*)target)->ApplyResistanceBuffModsMod(SpellSchools(x), m_positive, float(m_modifier.m_amount), apply);
-            }
-        }
-    }
-}
-
-/**
- * @brief Applies or removes percentage modifiers to base resistances.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleAuraModBaseResistancePCT(bool apply, bool /*Real*/)
-{
-    // only players have base stats
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        // pets only have base armor
-        if (((Creature*)GetTarget())->IsPet() && (m_modifier.m_miscvalue & SPELL_SCHOOL_MASK_NORMAL))
-        {
-            GetTarget()->HandleStatModifier(UNIT_MOD_ARMOR, BASE_PCT, float(m_modifier.m_amount), apply);
-        }
-    }
-    else
-    {
-        for (int8 x = SPELL_SCHOOL_NORMAL; x < MAX_SPELL_SCHOOL; ++x)
-        {
-            if (m_modifier.m_miscvalue & int32(1 << x))
-            {
-                GetTarget()->HandleStatModifier(UnitMods(UNIT_MOD_RESISTANCE_START + x), BASE_PCT, float(m_modifier.m_amount), apply);
-            }
-        }
-    }
-}
-
-/**
- * @brief Applies or removes percentage modifiers to total resistances.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleModResistancePercent(bool apply, bool /*Real*/)
-{
-    Unit* target = GetTarget();
-
-    for (int8 i = SPELL_SCHOOL_NORMAL; i < MAX_SPELL_SCHOOL; ++i)
-    {
-        if (m_modifier.m_miscvalue & int32(1 << i))
-        {
-            target->HandleStatModifier(UnitMods(UNIT_MOD_RESISTANCE_START + i), TOTAL_PCT, float(m_modifier.m_amount), apply);
-            if (target->GetTypeId() == TYPEID_PLAYER || ((Creature*)target)->IsPet())
-            {
-                ((Player*)target)->ApplyResistanceBuffModsPercentMod(SpellSchools(i), true, float(m_modifier.m_amount), apply);
-                ((Player*)target)->ApplyResistanceBuffModsPercentMod(SpellSchools(i), false, float(m_modifier.m_amount), apply);
-            }
-        }
-    }
-}
-
-/**
- * @brief Applies or removes flat modifiers to base resistances.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleModBaseResistance(bool apply, bool /*Real*/)
-{
-    // only players have base stats
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        // only pets have base stats
-        if (((Creature*)GetTarget())->IsPet() && (m_modifier.m_miscvalue & SPELL_SCHOOL_MASK_NORMAL))
-        {
-            GetTarget()->HandleStatModifier(UNIT_MOD_ARMOR, TOTAL_VALUE, float(m_modifier.m_amount), apply);
-        }
-    }
-    else
-    {
-        for (int i = SPELL_SCHOOL_NORMAL; i < MAX_SPELL_SCHOOL; ++i)
-        {
-            if (m_modifier.m_miscvalue & (1 << i))
-            {
-                GetTarget()->HandleStatModifier(UnitMods(UNIT_MOD_RESISTANCE_START + i), TOTAL_VALUE, float(m_modifier.m_amount), apply);
-            }
-        }
-    }
-}
-
 /********************************/
 /***           STAT           ***/
 /********************************/
 
-void Aura::HandleAuraModStat(bool apply, bool /*Real*/)
-{
-    if (m_modifier.m_miscvalue < -2 || m_modifier.m_miscvalue > 4)
-    {
-        sLog.outError("WARNING: Spell %u effect %u have unsupported misc value (%i) for SPELL_AURA_MOD_STAT ", GetId(), GetEffIndex(), m_modifier.m_miscvalue);
-        return;
-    }
-
-    for (int32 i = STAT_STRENGTH; i < MAX_STATS; ++i)
-    {
-        // -1 or -2 is all stats ( misc < -2 checked in function beginning )
-        if (m_modifier.m_miscvalue < 0 || m_modifier.m_miscvalue == i)
-        {
-            // m_target->ApplyStatMod(Stats(i), m_modifier.m_amount,apply);
-            GetTarget()->HandleStatModifier(UnitMods(UNIT_MOD_STAT_START + i), TOTAL_VALUE, float(m_modifier.m_amount), apply);
-            if (GetTarget()->GetTypeId() == TYPEID_PLAYER || ((Creature*)GetTarget())->IsPet())
-            {
-                ((Player*)GetTarget())->ApplyStatBuffMod(Stats(i), float(m_modifier.m_amount), apply);
-            }
-        }
-    }
-}
-
-/**
- * @brief Applies or removes percentage modifiers to player base stats.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleModPercentStat(bool apply, bool /*Real*/)
-{
-    if (m_modifier.m_miscvalue < -1 || m_modifier.m_miscvalue > 4)
-    {
-        sLog.outError("WARNING: Misc Value for SPELL_AURA_MOD_PERCENT_STAT not valid");
-        return;
-    }
-
-    // only players have base stats
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    for (int32 i = STAT_STRENGTH; i < MAX_STATS; ++i)
-    {
-        if (m_modifier.m_miscvalue == i || m_modifier.m_miscvalue == -1)
-        {
-            GetTarget()->HandleStatModifier(UnitMods(UNIT_MOD_STAT_START + i), BASE_PCT, float(m_modifier.m_amount), apply);
-        }
-    }
-}
-
-/**
- * @brief Refreshes player spell damage and healing data derived from stats.
- */
-void Aura::HandleModSpellDamagePercentFromStat(bool /*apply*/, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // Magic damage modifiers implemented in Unit::SpellDamageBonusDone
-    // This information for client side use only
-    // Recalculate bonus
-    ((Player*)GetTarget())->UpdateSpellDamageAndHealingBonus();
-}
-
-/**
- * @brief Refreshes player healing data derived from stats.
- */
-void Aura::HandleModSpellHealingPercentFromStat(bool /*apply*/, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // Recalculate bonus
-    ((Player*)GetTarget())->UpdateSpellDamageAndHealingBonus();
-}
-
-void Aura::HandleAuraModDispelResist(bool apply, bool Real)
-{
-    if (!Real || !apply)
-    {
-        return;
-    }
-
-    if (GetId() == 33206)
-    {
-        GetTarget()->CastSpell(GetTarget(), 44416, true, NULL, this, GetCasterGuid());
-    }
-}
-
-void Aura::HandleModSpellDamagePercentFromAttackPower(bool /*apply*/, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // Magic damage modifiers implemented in Unit::SpellDamageBonusDone
-    // This information for client side use only
-    // Recalculate bonus
-    ((Player*)GetTarget())->UpdateSpellDamageAndHealingBonus();
-}
-
-void Aura::HandleModSpellHealingPercentFromAttackPower(bool /*apply*/, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // Recalculate bonus
-    ((Player*)GetTarget())->UpdateSpellDamageAndHealingBonus();
-}
-
-/**
- * @brief Refreshes player healing bonus data exposed to the client.
- */
-void Aura::HandleModHealingDone(bool /*apply*/, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-    // implemented in Unit::SpellHealingBonusDone
-    // this information is for client side only
-    ((Player*)GetTarget())->UpdateSpellDamageAndHealingBonus();
-}
-
-/**
- * @brief Applies or removes percentage modifiers to total stats and preserves health ratios when required.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleModTotalPercentStat(bool apply, bool /*Real*/)
-{
-    if (m_modifier.m_miscvalue < -1 || m_modifier.m_miscvalue > 4)
-    {
-        sLog.outError("WARNING: Misc Value for SPELL_AURA_MOD_PERCENT_STAT not valid");
-        return;
-    }
-
-    Unit* target = GetTarget();
-
-    // save current and max HP before applying aura
-    uint32 curHPValue = target->GetHealth();
-    uint32 maxHPValue = target->GetMaxHealth();
-
-    for (int32 i = STAT_STRENGTH; i < MAX_STATS; ++i)
-    {
-        if (m_modifier.m_miscvalue == i || m_modifier.m_miscvalue == -1)
-        {
-            target->HandleStatModifier(UnitMods(UNIT_MOD_STAT_START + i), TOTAL_PCT, float(m_modifier.m_amount), apply);
-            if (target->GetTypeId() == TYPEID_PLAYER || ((Creature*)target)->IsPet())
-            {
-                ((Player*)target)->ApplyStatPercentBuffMod(Stats(i), float(m_modifier.m_amount), apply);
-            }
-        }
-    }
-
-    // recalculate current HP/MP after applying aura modifications (only for spells with 0x10 flag)
-    if (m_modifier.m_miscvalue == STAT_STAMINA && maxHPValue > 0 && GetSpellProto()->HasAttribute(SPELL_ATTR_ABILITY))
-    {
-        // newHP = (curHP / maxHP) * newMaxHP = (newMaxHP * curHP) / maxHP -> which is better because no int -> double -> int conversion is needed
-        uint32 newHPValue = (target->GetMaxHealth() * curHPValue) / maxHPValue;
-        target->SetHealth(newHPValue);
-    }
-}
-
-/**
- * @brief Refreshes armor-from-stat style resistance data for players.
- */
-void Aura::HandleAuraModResistenceOfStatPercent(bool /*apply*/, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (m_modifier.m_miscvalue != SPELL_SCHOOL_MASK_NORMAL)
-    {
-        // support required adding replace UpdateArmor by loop by UpdateResistence at intellect update
-        // and include in UpdateResistence same code as in UpdateArmor for aura mod apply.
-        sLog.outError("Aura SPELL_AURA_MOD_RESISTANCE_OF_STAT_PERCENT(182) need adding support for non-armor resistances!");
-        return;
-    }
-
-    // Recalculate Armor
-    GetTarget()->UpdateArmor();
-}
-
-/********************************/
-/***      HEAL & ENERGIZE     ***/
-/********************************/
-void Aura::HandleAuraModTotalHealthPercentRegen(bool apply, bool /*Real*/)
-{
-    m_isPeriodic = apply;
-}
-
-/**
- * @brief Enables periodic total mana regeneration using a one-second tick when needed.
- *
- * @param apply True to enable the periodic effect; false to disable it.
- */
-void Aura::HandleAuraModTotalManaPercentRegen(bool apply, bool /*Real*/)
-{
-    if (m_modifier.periodictime == 0)
-    {
-        m_modifier.periodictime = 1000;
-    }
-
-    m_periodicTimer = m_modifier.periodictime;
-    m_isPeriodic = apply;
-}
-
-void Aura::HandleModRegen(bool apply, bool /*Real*/)        // eating
-{
-    if (m_modifier.periodictime == 0)
-    {
-        m_modifier.periodictime = 5000;
-    }
-
-    m_periodicTimer = 5000;
-    m_isPeriodic = apply;
-}
-
-void Aura::HandleModPowerRegen(bool apply, bool Real)       // drinking
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    Powers powerType = GetTarget()->GetPowerType();
-    if (m_modifier.periodictime == 0)
-    {
-        // Anger Management (only spell use this aura for rage)
-        if (powerType == POWER_RAGE)
-        {
-            m_modifier.periodictime = 3000;
-        }
-        else
-        {
-            m_modifier.periodictime = 2000;
-        }
-    }
-
-    m_periodicTimer = 5000;
-
-    if (GetTarget()->GetTypeId() == TYPEID_PLAYER && m_modifier.m_miscvalue == POWER_MANA)
-    {
-        ((Player*)GetTarget())->UpdateManaRegen();
-    }
-
-    m_isPeriodic = apply;
-}
-
-/**
- * @brief Refreshes player mana regeneration after percentage-based regen changes.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleModPowerRegenPCT(bool /*apply*/, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // Update manaregen value
-    if (m_modifier.m_miscvalue == POWER_MANA)
-    {
-        ((Player*)GetTarget())->UpdateManaRegen();
-    }
-}
-
-void Aura::HandleModManaRegen(bool /*apply*/, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // Note: an increase in regen does NOT cause threat.
-    ((Player*)GetTarget())->UpdateManaRegen();
-}
-
-void Aura::HandleComprehendLanguage(bool apply, bool /*Real*/)
-{
-    if (apply)
-    {
-        GetTarget()->SetFlag(UNIT_FIELD_FLAGS_2, UNIT_FLAG2_COMPREHEND_LANG);
-    }
-    else
-    {
-        GetTarget()->RemoveFlag(UNIT_FIELD_FLAGS_2, UNIT_FLAG2_COMPREHEND_LANG);
-    }
-}
-
-/**
- * @brief Applies or removes maximum health increases, including special temporary-health cases.
- *
- * @param apply True to apply the modifier; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraModIncreaseHealth(bool apply, bool Real)
-{
-    Unit* target = GetTarget();
-
-    switch (GetId())
-    {
-            // Special case with temporary increase max/current health
-            // Cases where we need to manually calculate the amount for the spell (by percentage)
-            // recalculate to full amount at apply for proper remove
-            // Backport notive TBC: no cases yet
-            // no break here
-
-            // Cases where m_amount already has the correct value (spells cast with CastCustomSpell or absolute values)
-        case 12976:                                         // Warrior Last Stand triggered spell (Cast with percentage-value by CastCustomSpell)
-        case 28726:                                         // Nightmare Seed
-        case 34511:                                         // Valor (Bulwark of Kings, Bulwark of the Ancient Kings)
-        case 44055:                                         // Tremendous Fortitude (Battlemaster's Alacrity)
-        {
-            if (Real)
-            {
-                if (apply)
-                {
-                    target->HandleStatModifier(UNIT_MOD_HEALTH, TOTAL_VALUE, float(m_modifier.m_amount), apply);
-                    target->ModifyHealth(m_modifier.m_amount);
-                }
-                else
-                {
-                    if (int32(target->GetHealth()) > m_modifier.m_amount)
-                    {
-                        target->ModifyHealth(-m_modifier.m_amount);
-                    }
-                    else
-                    {
-                        target->SetHealth(1);
-                    }
-                    target->HandleStatModifier(UNIT_MOD_HEALTH, TOTAL_VALUE, float(m_modifier.m_amount), apply);
-                }
-            }
-            return;
-        }
-        // generic case
-        default:
-            target->HandleStatModifier(UNIT_MOD_HEALTH, TOTAL_VALUE, float(m_modifier.m_amount), apply);
-    }
-}
-
-void  Aura::HandleAuraModIncreaseMaxHealth(bool apply, bool /*Real*/)
-{
-    Unit* target = GetTarget();
-    uint32 oldhealth = target->GetHealth();
-    double healthPercentage = (double)oldhealth / (double)target->GetMaxHealth();
-
-    target->HandleStatModifier(UNIT_MOD_HEALTH, TOTAL_VALUE, float(m_modifier.m_amount), apply);
-
-    // refresh percentage
-    if (oldhealth > 0)
-    {
-        uint32 newhealth = uint32(ceil((double)target->GetMaxHealth() * healthPercentage));
-        if (newhealth == 0)
-        {
-            newhealth = 1;
-        }
-
-        target->SetHealth(newhealth);
-    }
-}
-
-/**
- * @brief Applies or removes a flat increase to the current power type's maximum value.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleAuraModIncreaseEnergy(bool apply, bool /*Real*/)
-{
-    Unit* target = GetTarget();
-    Powers powerType = target->GetPowerType();
-    if (int32(powerType) != m_modifier.m_miscvalue)
-    {
-        return;
-    }
-
-    UnitMods unitMod = UnitMods(UNIT_MOD_POWER_START + powerType);
-
-    target->HandleStatModifier(unitMod, TOTAL_VALUE, float(m_modifier.m_amount), apply);
-}
-
-/**
- * @brief Applies or removes a percentage increase to the current power type.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleAuraModIncreaseEnergyPercent(bool apply, bool /*Real*/)
-{
-    Powers powerType = GetTarget()->GetPowerType();
-    if (int32(powerType) != m_modifier.m_miscvalue)
-    {
-        return;
-    }
-
-    UnitMods unitMod = UnitMods(UNIT_MOD_POWER_START + powerType);
-
-    GetTarget()->HandleStatModifier(unitMod, TOTAL_PCT, float(m_modifier.m_amount), apply);
-}
-
-/**
- * @brief Applies or removes a percentage increase to maximum health.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleAuraModIncreaseHealthPercent(bool apply, bool /*Real*/)
-{
-    GetTarget()->HandleStatModifier(UNIT_MOD_HEALTH, TOTAL_PCT, float(m_modifier.m_amount), apply);
-}
-
 /********************************/
 /***          FIGHT           ***/
 /********************************/
-
-void Aura::HandleAuraModParryPercent(bool /*apply*/, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)GetTarget())->UpdateParryPercentage();
-}
-
-/**
- * @brief Refreshes player dodge percentage after aura changes.
- */
-void Aura::HandleAuraModDodgePercent(bool /*apply*/, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)GetTarget())->UpdateDodgePercentage();
-    // sLog.outError("BONUS DODGE CHANCE: + %f", float(m_modifier.m_amount));
-}
 
 /**
  * @brief Refreshes player block percentage after aura changes.
@@ -4063,385 +2210,13 @@ void Aura::HandleAuraModBlockPercent(bool /*apply*/, bool /*Real*/)
     // sLog.outError("BONUS BLOCK CHANCE: + %f", float(m_modifier.m_amount));
 }
 
-/**
- * @brief Refreshes player mana regeneration rules after regen-interrupt changes.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraModRegenInterrupt(bool /*apply*/, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)GetTarget())->UpdateManaRegen();
-}
-
-/**
- * @brief Applies or removes melee and ranged critical strike bonuses.
- *
- * @param apply True to apply the modifier; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleAuraModCritPercent(bool apply, bool Real)
-{
-    Unit* target = GetTarget();
-
-    if (target->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // apply item specific bonuses for already equipped weapon
-    if (Real)
-    {
-        for (int i = 0; i < MAX_ATTACK; ++i)
-        {
-            if (Item* pItem = ((Player*)target)->GetWeaponForAttack(WeaponAttackType(i), true, false))
-            {
-                ((Player*)target)->_ApplyWeaponDependentAuraCritMod(pItem, WeaponAttackType(i), this, apply);
-            }
-        }
-    }
-
-    // mods must be applied base at equipped weapon class and subclass comparison
-    // with spell->EquippedItemClass and  EquippedItemSubClassMask and EquippedItemInventoryTypeMask
-    // m_modifier.m_miscvalue comparison with item generated damage types
-
-    if (GetSpellProto()->EquippedItemClass == -1)
-    {
-        ((Player*)target)->HandleBaseModValue(CRIT_PERCENTAGE,         FLAT_MOD, float(m_modifier.m_amount), apply);
-        ((Player*)target)->HandleBaseModValue(OFFHAND_CRIT_PERCENTAGE, FLAT_MOD, float(m_modifier.m_amount), apply);
-        ((Player*)target)->HandleBaseModValue(RANGED_CRIT_PERCENTAGE,  FLAT_MOD, float(m_modifier.m_amount), apply);
-    }
-    else
-    {
-        // done in Player::_ApplyWeaponDependentAuraMods
-    }
-}
-
-/**
- * @brief Applies or removes melee and ranged hit chance modifiers.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleModHitChance(bool apply, bool /*Real*/)
-{
-    Unit* target = GetTarget();
-
-    if (target->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)target)->UpdateMeleeHitChances();
-        ((Player*)target)->UpdateRangedHitChances();
-    }
-    else
-    {
-        target->m_modMeleeHitChance += apply ? m_modifier.m_amount : (-m_modifier.m_amount);
-        target->m_modRangedHitChance += apply ? m_modifier.m_amount : (-m_modifier.m_amount);
-    }
-}
-
-/**
- * @brief Applies or removes spell hit chance modifiers.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleModSpellHitChance(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)GetTarget())->UpdateSpellHitChances();
-    }
-    else
-    {
-        GetTarget()->m_modSpellHitChance += apply ? m_modifier.m_amount : (-m_modifier.m_amount);
-    }
-}
-
-/**
- * @brief Applies or removes spell critical strike chance bonuses.
- *
- * @param apply True to apply the modifier; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleModSpellCritChance(bool apply, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    if (GetTarget()->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)GetTarget())->UpdateAllSpellCritChances();
-    }
-    else
-    {
-        GetTarget()->m_baseSpellCritChance += apply ? m_modifier.m_amount : (-m_modifier.m_amount);
-    }
-}
-
-/**
- * @brief Refreshes per-school spell critical strike chance for players.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleModSpellCritChanceShool(bool /*apply*/, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    for (int school = SPELL_SCHOOL_NORMAL; school < MAX_SPELL_SCHOOL; ++school)
-    {
-        if (m_modifier.m_miscvalue & (1 << school))
-        {
-            ((Player*)GetTarget())->UpdateSpellCritChance(school);
-        }
-    }
-}
-
 /********************************/
 /***         ATTACK SPEED     ***/
 /********************************/
 
-void Aura::HandleModCastingSpeed(bool apply, bool /*Real*/)
-{
-    if (apply)
-    {
-        if (Unit* caster = GetCaster())
-        {
-            if (Player* modOwner = caster->GetSpellModOwner())
-            {
-                modOwner->ApplySpellMod(GetSpellProto()->Id, SPELLMOD_HASTE, m_modifier.m_amount);
-            }
-        }
-    }
-
-    GetTarget()->ApplyCastTimePercentMod(m_modifier.m_amount, apply);
-}
-
-void Aura::HandleModMeleeRangedSpeedPct(bool apply, bool /*Real*/)
-{
-    Unit* target = GetTarget();
-    target->ApplyAttackTimePercentMod(BASE_ATTACK, float(m_modifier.m_amount), apply);
-    target->ApplyAttackTimePercentMod(OFF_ATTACK, float(m_modifier.m_amount), apply);
-    target->ApplyAttackTimePercentMod(RANGED_ATTACK, float(m_modifier.m_amount), apply);
-}
-
-void Aura::HandleModCombatSpeedPct(bool apply, bool /*Real*/)
-{
-    Unit* target = GetTarget();
-    target->ApplyCastTimePercentMod(float(m_modifier.m_amount), apply);
-    target->ApplyAttackTimePercentMod(BASE_ATTACK, float(m_modifier.m_amount), apply);
-    target->ApplyAttackTimePercentMod(OFF_ATTACK, float(m_modifier.m_amount), apply);
-    target->ApplyAttackTimePercentMod(RANGED_ATTACK, float(m_modifier.m_amount), apply);
-}
-
-/**
- * @brief Applies or removes haste effects to main-hand attack speed.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleModAttackSpeed(bool apply, bool /*Real*/)
-{
-    if (apply)
-    {
-        if (Unit* caster = GetCaster())
-        {
-            if (Player* modOwner = caster->GetSpellModOwner())
-            {
-                modOwner->ApplySpellMod(GetSpellProto()->Id, SPELLMOD_HASTE, m_modifier.m_amount);
-            }
-        }
-    }
-
-    GetTarget()->ApplyAttackTimePercentMod(BASE_ATTACK, m_modifier.m_amount, apply);
-}
-
-/**
- * @brief Applies or removes haste effects to both melee attack speeds.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleModMeleeSpeedPct(bool apply, bool /*Real*/)
-{
-    if (apply)
-    {
-        if (Unit* caster = GetCaster())
-        {
-            if (Player* modOwner = caster->GetSpellModOwner())
-            {
-                modOwner->ApplySpellMod(GetSpellProto()->Id, SPELLMOD_HASTE, m_modifier.m_amount);
-            }
-        }
-    }
-
-    Unit* target = GetTarget();
-    target->ApplyAttackTimePercentMod(BASE_ATTACK, m_modifier.m_amount, apply);
-    target->ApplyAttackTimePercentMod(OFF_ATTACK, m_modifier.m_amount, apply);
-}
-
-/**
- * @brief Applies or removes haste effects to ranged attack speed.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleAuraModRangedHaste(bool apply, bool /*Real*/)
-{
-    if (apply)
-    {
-        if (Unit* caster = GetCaster())
-        {
-            if (Player* modOwner = caster->GetSpellModOwner())
-            {
-                modOwner->ApplySpellMod(GetSpellProto()->Id, SPELLMOD_HASTE, m_modifier.m_amount);
-            }
-        }
-    }
-
-    GetTarget()->ApplyAttackTimePercentMod(RANGED_ATTACK, m_modifier.m_amount, apply);
-}
-
-/**
- * @brief Applies or removes ammo-based ranged haste when the equipped weapon uses ammunition.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleRangedAmmoHaste(bool apply, bool /*Real*/)
-{
-    if (GetTarget()->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // Quivers should not increase attack speed for ranged weapons which do not require any ammo.
-    Item* ranged_weapon = GetTarget()->ToPlayer()->GetItemByPos(INVENTORY_SLOT_BAG_0, EQUIPMENT_SLOT_RANGED);
-    if (!ranged_weapon || ranged_weapon->GetProto()->AmmoType == 0)
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        if (Unit* caster = GetCaster())
-        {
-            if (Player* modOwner = caster->GetSpellModOwner())
-            {
-                modOwner->ApplySpellMod(GetSpellProto()->Id, SPELLMOD_HASTE, m_modifier.m_amount);
-            }
-        }
-    }
-
-    GetTarget()->ApplyAttackTimePercentMod(RANGED_ATTACK, m_modifier.m_amount, apply);
-}
-
 /********************************/
 /***        ATTACK POWER      ***/
 /********************************/
-
-void Aura::HandleAuraModAttackPower(bool apply, bool /*Real*/)
-{
-    if (apply)
-    {
-        if (Unit* caster = GetCaster())
-        {
-            if (Player* modOwner = caster->GetSpellModOwner())
-            {
-                modOwner->ApplySpellMod(GetSpellProto()->Id, SPELLMOD_ATTACK_POWER, m_modifier.m_amount);
-            }
-        }
-    }
-
-    GetTarget()->HandleStatModifier(UNIT_MOD_ATTACK_POWER, TOTAL_VALUE, m_modifier.m_amount, apply);
-}
-
-/**
- * @brief Applies or removes flat ranged attack power bonuses.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleAuraModRangedAttackPower(bool apply, bool /*Real*/)
-{
-    if ((GetTarget()->getClassMask() & CLASSMASK_WAND_USERS) != 0)
-    {
-        return;
-    }
-
-    if (apply)
-    {
-        if (Unit* caster = GetCaster())
-        {
-            if (Player* modOwner = caster->GetSpellModOwner())
-            {
-                modOwner->ApplySpellMod(GetSpellProto()->Id, SPELLMOD_ATTACK_POWER, m_modifier.m_amount);
-            }
-        }
-    }
-
-    GetTarget()->HandleStatModifier(UNIT_MOD_ATTACK_POWER_RANGED, TOTAL_VALUE, m_modifier.m_amount, apply);
-}
-
-/**
- * @brief Applies or removes percentage modifiers to melee attack power.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleAuraModAttackPowerPercent(bool apply, bool /*Real*/)
-{
-    if (apply)
-    {
-        if (Unit* caster = GetCaster())
-        {
-            if (Player* modOwner = caster->GetSpellModOwner())
-            {
-                modOwner->ApplySpellMod(GetSpellProto()->Id, SPELLMOD_ATTACK_POWER, m_modifier.m_amount);
-            }
-        }
-    }
-
-    // UNIT_FIELD_ATTACK_POWER_MULTIPLIER = multiplier - 1
-    GetTarget()->HandleStatModifier(UNIT_MOD_ATTACK_POWER, TOTAL_PCT, m_modifier.m_amount, apply);
-}
-
-/**
- * @brief Applies or removes percentage modifiers to ranged attack power.
- *
- * @param apply True to apply the modifier; false to remove it.
- */
-void Aura::HandleAuraModRangedAttackPowerPercent(bool apply, bool /*Real*/)
-{
-    if ((GetTarget()->getClassMask() & CLASSMASK_WAND_USERS) != 0)
-    {
-        return;
-    }
-
-    float amount = m_modifier.m_amount;
-
-    if (Unit* caster = GetCaster())
-    {
-        if (Player* modOwner = caster->GetSpellModOwner())
-        {
-            modOwner->ApplySpellMod(GetSpellProto()->Id, SPELLMOD_ATTACK_POWER, amount);
-        }
-    }
-
-    // UNIT_FIELD_RANGED_ATTACK_POWER_MULTIPLIER = multiplier - 1
-    GetTarget()->HandleStatModifier(UNIT_MOD_ATTACK_POWER_RANGED, TOTAL_PCT, amount, apply);
-}
 
 void Aura::HandleAuraModRangedAttackPowerOfStatPercent(bool /*apply*/, bool Real)
 {
@@ -4459,246 +2234,8 @@ void Aura::HandleAuraModRangedAttackPowerOfStatPercent(bool /*apply*/, bool Real
 }
 
 /********************************/
-/***        DAMAGE BONUS      ***/
-/********************************/
-void Aura::HandleModDamageDone(bool apply, bool Real)
-{
-    Unit* target = GetTarget();
-
-    // apply item specific bonuses for already equipped weapon
-    if (Real && target->GetTypeId() == TYPEID_PLAYER)
-    {
-        for (int i = 0; i < MAX_ATTACK; ++i)
-        {
-            if (Item* pItem = ((Player*)target)->GetWeaponForAttack(WeaponAttackType(i), true, false))
-            {
-                ((Player*)target)->_ApplyWeaponDependentAuraDamageMod(pItem, WeaponAttackType(i), this, apply);
-            }
-        }
-    }
-
-    // m_modifier.m_miscvalue is bitmask of spell schools
-    // 1 ( 0-bit ) - normal school damage (SPELL_SCHOOL_MASK_NORMAL)
-    // 126 - full bitmask all magic damages (SPELL_SCHOOL_MASK_MAGIC) including wands
-    // 127 - full bitmask any damages
-    //
-    // mods must be applied base at equipped weapon class and subclass comparison
-    // with spell->EquippedItemClass and  EquippedItemSubClassMask and EquippedItemInventoryTypeMask
-    // m_modifier.m_miscvalue comparison with item generated damage types
-
-    if ((m_modifier.m_miscvalue & SPELL_SCHOOL_MASK_NORMAL) != 0)
-    {
-        // apply generic physical damage bonuses including wand case
-        if (GetSpellProto()->EquippedItemClass == -1 || target->GetTypeId() != TYPEID_PLAYER)
-        {
-            target->HandleStatModifier(UNIT_MOD_DAMAGE_MAINHAND, TOTAL_VALUE, float(m_modifier.m_amount), apply);
-            target->HandleStatModifier(UNIT_MOD_DAMAGE_OFFHAND, TOTAL_VALUE, float(m_modifier.m_amount), apply);
-            target->HandleStatModifier(UNIT_MOD_DAMAGE_RANGED, TOTAL_VALUE, float(m_modifier.m_amount), apply);
-        }
-        else
-        {
-            // done in Player::_ApplyWeaponDependentAuraMods
-        }
-
-        if (target->GetTypeId() == TYPEID_PLAYER)
-        {
-            if (m_positive)
-            {
-                target->ApplyModUInt32Value(PLAYER_FIELD_MOD_DAMAGE_DONE_POS, m_modifier.m_amount, apply);
-            }
-            else
-            {
-                target->ApplyModUInt32Value(PLAYER_FIELD_MOD_DAMAGE_DONE_NEG, m_modifier.m_amount, apply);
-            }
-        }
-    }
-
-    // Skip non magic case for speedup
-    if ((m_modifier.m_miscvalue & SPELL_SCHOOL_MASK_MAGIC) == 0)
-    {
-        return;
-    }
-
-    if (GetSpellProto()->EquippedItemClass != -1 || GetSpellProto()->EquippedItemInventoryTypeMask != 0)
-    {
-        // wand magic case (skip generic to all item spell bonuses)
-        // done in Player::_ApplyWeaponDependentAuraMods
-
-        // Skip item specific requirements for not wand magic damage
-        return;
-    }
-
-    // Magic damage modifiers implemented in Unit::SpellDamageBonusDone
-    // This information for client side use only
-    if (target->GetTypeId() == TYPEID_PLAYER)
-    {
-        if (m_positive)
-        {
-            for (int i = SPELL_SCHOOL_HOLY; i < MAX_SPELL_SCHOOL; ++i)
-            {
-                if ((m_modifier.m_miscvalue & (1 << i)) != 0)
-                {
-                    target->ApplyModUInt32Value(PLAYER_FIELD_MOD_DAMAGE_DONE_POS + i, m_modifier.m_amount, apply);
-                }
-            }
-        }
-        else
-        {
-            for (int i = SPELL_SCHOOL_HOLY; i < MAX_SPELL_SCHOOL; ++i)
-            {
-                if ((m_modifier.m_miscvalue & (1 << i)) != 0)
-                {
-                    target->ApplyModUInt32Value(PLAYER_FIELD_MOD_DAMAGE_DONE_NEG + i, m_modifier.m_amount, apply);
-                }
-            }
-        }
-        Pet* pet = target->GetPet();
-        if (pet)
-        {
-            pet->UpdateAttackPowerAndDamage();
-        }
-    }
-}
-
-/**
- * @brief Applies or removes percentage damage bonuses for physical and magical damage.
- *
- * @param apply True to apply the modifier; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleModDamagePercentDone(bool apply, bool Real)
-{
-    DEBUG_FILTER_LOG(LOG_FILTER_SPELL_CAST, "AURA MOD DAMAGE type:%u negative:%u", m_modifier.m_miscvalue, m_positive ? 0 : 1);
-    Unit* target = GetTarget();
-
-    // apply item specific bonuses for already equipped weapon
-    if (Real && target->GetTypeId() == TYPEID_PLAYER)
-    {
-        for (int i = 0; i < MAX_ATTACK; ++i)
-        {
-            if (Item* pItem = ((Player*)target)->GetWeaponForAttack(WeaponAttackType(i), true, false))
-            {
-                ((Player*)target)->_ApplyWeaponDependentAuraDamageMod(pItem, WeaponAttackType(i), this, apply);
-            }
-        }
-    }
-
-    // m_modifier.m_miscvalue is bitmask of spell schools
-    // 1 ( 0-bit ) - normal school damage (SPELL_SCHOOL_MASK_NORMAL)
-    // 126 - full bitmask all magic damages (SPELL_SCHOOL_MASK_MAGIC) including wand
-    // 127 - full bitmask any damages
-    //
-    // mods must be applied base at equipped weapon class and subclass comparison
-    // with spell->EquippedItemClass and  EquippedItemSubClassMask and EquippedItemInventoryTypeMask
-    // m_modifier.m_miscvalue comparison with item generated damage types
-
-    if ((m_modifier.m_miscvalue & SPELL_SCHOOL_MASK_NORMAL) != 0)
-    {
-        // apply generic physical damage bonuses including wand case
-        if (GetSpellProto()->EquippedItemClass == -1 || target->GetTypeId() != TYPEID_PLAYER)
-        {
-            target->HandleStatModifier(UNIT_MOD_DAMAGE_MAINHAND, TOTAL_PCT, float(m_modifier.m_amount), apply);
-            target->HandleStatModifier(UNIT_MOD_DAMAGE_OFFHAND, TOTAL_PCT, float(m_modifier.m_amount), apply);
-            target->HandleStatModifier(UNIT_MOD_DAMAGE_RANGED, TOTAL_PCT, float(m_modifier.m_amount), apply);
-        }
-        else
-        {
-            // done in Player::_ApplyWeaponDependentAuraMods
-        }
-        // For show in client
-        if (target->GetTypeId() == TYPEID_PLAYER)
-        {
-            target->ApplyModSignedFloatValue(PLAYER_FIELD_MOD_DAMAGE_DONE_PCT, m_modifier.m_amount / 100.0f, apply);
-        }
-    }
-
-    // Skip non magic case for speedup
-    if ((m_modifier.m_miscvalue & SPELL_SCHOOL_MASK_MAGIC) == 0)
-    {
-        return;
-    }
-
-    if (GetSpellProto()->EquippedItemClass != -1 || GetSpellProto()->EquippedItemInventoryTypeMask != 0)
-    {
-        // wand magic case (skip generic to all item spell bonuses)
-        // done in Player::_ApplyWeaponDependentAuraMods
-
-        // Skip item specific requirements for not wand magic damage
-        return;
-    }
-
-    // Magic damage percent modifiers implemented in Unit::SpellDamageBonusDone
-    // Send info to client
-    if (target->GetTypeId() == TYPEID_PLAYER)
-        for (int i = SPELL_SCHOOL_HOLY; i < MAX_SPELL_SCHOOL; ++i)
-        {
-            target->ApplyModSignedFloatValue(PLAYER_FIELD_MOD_DAMAGE_DONE_PCT + i, m_modifier.m_amount / 100.0f, apply);
-        }
-}
-
-/**
- * @brief Applies or removes percentage damage bonuses to offhand attacks.
- *
- * @param apply True to apply the modifier; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleModOffhandDamagePercent(bool apply, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    DEBUG_FILTER_LOG(LOG_FILTER_SPELL_CAST, "AURA MOD OFFHAND DAMAGE");
-
-    GetTarget()->HandleStatModifier(UNIT_MOD_DAMAGE_OFFHAND, TOTAL_PCT, float(m_modifier.m_amount), apply);
-}
-
-/********************************/
 /***        POWER COST        ***/
 /********************************/
-
-void Aura::HandleModPowerCostPCT(bool apply, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    float amount = m_modifier.m_amount / 100.0f;
-    for (int i = 0; i < MAX_SPELL_SCHOOL; ++i)
-    {
-        if (m_modifier.m_miscvalue & (1 << i))
-        {
-            GetTarget()->ApplyModSignedFloatValue(UNIT_FIELD_POWER_COST_MULTIPLIER + i, amount, apply);
-        }
-    }
-}
-
-/**
- * @brief Applies or removes flat spell power cost modifiers by school.
- *
- * @param apply True to apply the modifier; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleModPowerCost(bool apply, bool Real)
-{
-    // spells required only Real aura add/remove
-    if (!Real)
-    {
-        return;
-    }
-
-    for (int i = 0; i < MAX_SPELL_SCHOOL; ++i)
-    {
-        if (m_modifier.m_miscvalue & (1 << i))
-        {
-            GetTarget()->ApplyModInt32Value(UNIT_FIELD_POWER_COST_MODIFIER + i, m_modifier.m_amount, apply);
-        }
-    }
-}
 
 /*********************************************************/
 /***                    OTHERS                         ***/
